Move ContentView into the layout folder
